Separate the authActions from the user actions
